authentication created

+++ src/modules/Admin/admin.service.ts
@@ -1,3 +1,5 @@
+import { StatusCodes } from "http-status-codes";
+import AppError from "../../error/AppError";
 import { Blog } from "../Blog/blog.model";
 import { TUser } from "../User/user.interface";
 import { User } from "../User/user.model";
@@ -11,6 +13,10 @@ const blockUserInDB = async (userId: string, payload: Partial<TUser>) => {
 };
 
 const deleteBlogsFromDB = async (id: string) => {
+  if (!id) {
+    throw new AppError("Blog not found", StatusCodes.NOT_FOUND);
+  }
+
   const blogs = await Blog.findByIdAndDelete(id);
   return blogs;
 };
